+++ config/webpack.base.js
@@ -2,14 +2,17 @@
 /* eslint-disable import/no-extraneous-dependencies */
 
 const path = require('path');
+const webpack = require('webpack');
 const autoprefixer = require('autoprefixer');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 module.exports = {
     entry: './src/index.js',
 
+    context: path.resolve(__dirname, '..'),
+
     output: {
-        filename: 'bundle.js',
+        filename: 'gypcrete.js',
         path: path.resolve(__dirname, '../dist'),
         library: 'gypcrete'
     },
@@ -19,8 +22,7 @@ module.exports = {
             {
                 test: /\.jsx?$/,
                 include: [
-                    path.resolve(__dirname, '../src'),
-                    path.resolve(__dirname, '../doc')
+                    path.resolve(__dirname, '../src')
                 ],
                 use: ['babel-loader']
             },
@@ -52,11 +54,28 @@ module.exports = {
                         }
                     ]
                 })
+            },
+            {
+                test: /\.(woff|woff2|otf|ttf|eot|svg)$/,
+                include: [
+                    path.resolve(__dirname, '../src/fonts')
+                ],
+                use: [
+                    {
+                        loader: 'file-loader',
+                        options: {
+                            name: '[name]-[hash:6].[ext]',
+                            outputPath: 'fonts/',
+                            publicPath: 'fonts/'
+                        }
+                    }
+                ]
             }
         ]
     },
 
     plugins: [
-        new ExtractTextPlugin('bundle.css')
+        new webpack.ProgressPlugin(),
+        new ExtractTextPlugin('gypcrete.css')
     ]
 };