+++ demo/RowComponents/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import TextLabelDoc from './TextLabelDoc';
+
+function RowComponents() {
+    return (
+        <div>
+            <h1>Row Components</h1>
+
+            <TextLabelDoc />
+        </div>
+    );
+}
+
+export default RowComponents;