+++ demo/VisualElements/TagDoc.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Tag from 'src/Tag';
+
+function TagDoc() {
+    return (
+        <div>
+            <h2>&lt;Tag&gt;</h2>
+
+            <Tag>Tag</Tag>
+        </div>
+    );
+}
+
+export default TagDoc;