+++ demo/VisualElements/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+import TextDoc from './TextDoc';
+import IconDoc from './IconDoc';
+import TagDoc from './TagDoc';
+
+function VisualElements() {
+    return (
+        <div>
+            <h1>Visual Elements</h1>
+
+            <TextDoc />
+            <IconDoc />
+            <TagDoc />
+        </div>
+    );
+}
+
+export default VisualElements;