+++ doc/DocApp.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import App from '../src';
-
-function DocApp() {
-    return <App />;
-}
-
-export default DocApp;