+++ src/FlexCell.js
@@ -1,0 +1,40 @@
+import React, { PropTypes } from 'react';
+import icBEM from './utils/icBEM';
+import './styles/FlexCell.scss';
+
+const COMPONENT_NAME = 'ic-flex-cell';
+const ROOT_BEM = icBEM(COMPONENT_NAME);
+
+function FlexCell({ shrink, grow, basis, children }) {
+    const flexStyles = {
+        flexShrink: Number(shrink),
+        flexGrow: Number(grow),
+        flexBasis: basis,
+    };
+
+    return (
+        <div className={ROOT_BEM} style={flexStyles}>
+            {children}
+        </div>
+    );
+}
+
+FlexCell.propTypes = {
+    shrink: PropTypes.oneOfType([
+        PropTypes.bool,
+        PropTypes.number
+    ]),
+    grow: PropTypes.oneOfType([
+        PropTypes.bool,
+        PropTypes.number
+    ]),
+    basis: PropTypes.string,
+};
+
+FlexCell.defaultProps = {
+    shrink: false,
+    grow: false,
+    basis: 'auto',
+};
+
+export default FlexCell;