+++ src/__tests__/Icon.test.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '../index';
+import Icon from '../Icon';
 
 it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<App />, div);
+    const element = <Icon type="duplicate" />;
+
+    ReactDOM.render(element, div);
 });