+++ src/__tests__/RowCompBody.test.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import RowCompBody from '../RowCompBody';
+
+it('renders without crashing', () => {
+    const div = document.createElement('div');
+    const element = (
+        <RowCompBody>
+            test
+        </RowCompBody>
+    );
+
+    ReactDOM.render(element, div);
+});